integrate crypto.php into WC_Flexi_Gateway.php
